Changed method parameter to double from int.

+++ src/academy/learnprogramming/Main.java
@@ -7,23 +7,23 @@ public class Main {
         calcFeetAndInchesToCentimeters(71);
     }
 
-    public static int calcFeetAndInchesToCentimeters(int feet, int inches) {
+    public static double calcFeetAndInchesToCentimeters(double feet, double inches) {
         if (feet < 0 || (inches < 0 || inches > 12)) {
             return -1;
         } else {
             double centimeters = ((feet * 30.48) + (inches * 2.54));
             System.out.println(feet + "ft, " + inches + "in = " + centimeters + "cm");
-            return (int) centimeters;
+            return centimeters;
         }
     }
 
-    public static int calcFeetAndInchesToCentimeters(int inches) {
+    public static double calcFeetAndInchesToCentimeters(double inches) {
         if ((inches < 0)) {
             return -1;
         } else {
             double centimeters = ((inches * 2.54));
             System.out.println(inches + "in = " + centimeters + "cm");
-            return (int) centimeters;
+            return centimeters;
         }
     }
 }
